Create changeLog like a separete component and add loader
