+++ src/models/changeLog/selectors.ts
@@ -4,6 +4,6 @@ export const selectSize = (state: RootStateType): number | undefined => {
   return state.changesLog.size;
 };
 
-export const selectTotalPages = (state: RootStateType): number | undefined => {
+export const selectTotalPages = (state: RootStateType): number | unknown => {
   return state.changesLog.totalPages;
 };
